sensors/requirements.txt
Adafruit-BMP==1.5.4
Adafruit-GPIO==1.0.4
Adafruit-PureIO==1.1.8
dnspython==2.1.0
mysqlclient==2.0.3
pymongo==3.11.2
python-decouple==3.4
spidev==3.5
